Silence some -Wcast-function-type warnings
We choose to think that unused extra parameters to a function are
safely ignored, so that we can safely use a `void(*)(void*)` and call
it as `void(*)(void*, void*)` and it'll Do the Right Thing™.  However,
the compiler might easily warn about that, so hush it up.

Although kind of shady, this should work on any platform where GTK
does, so we should be fine.