Revert "Don't call sidebar_update_tag_list() and document_highlight_tags() on startup"
This detection based on the currently tab switch is imperfect as it
doesn't happen if there's a single document for instance so this
approach doesn't work and may lead to missed highlight requests.

Instead, introduce a new signal detecting session file opening
(in a separate PR) and use that one.

This reverts commit 952c4bf5e2522aa73d810b3ae9523bf36fc8dcdf.