encodings: Add a flag for whether the encoding is supported
This adds support for checking whether an encoding is actually
supported on the system, as some are commonly not supported everywhere,
e.g. ISO-IR-111, ISO-8859-8-I or HZ.

Here we assume that a conversion to UTF-8 means it also works the other
way around (e.g. from UTF-8, assuming the data can be represented, of
course).  This is a valid assumption if GNU libc or GNU libiconv is
used, and might or might not hold for other implementations.
This assumption however makes the code a fair bit simpler, and should
have little impact in reality as this flag is mostly a suggestion that
can be used to hide UI elements or skip attempts, and that in most
situations it doesn't make much sense for us to be able to *write* to
an encoding we cannot read.

As for startup performances, checking all encodings takes about 10ms on
my machine, which sounds Good Enough™.

This is currently unused, but subsequent commits will start using this.