CI: run tests inside an Xvfb X server
I opted for an X based solution with xvfb-run as that can easily be
installed via the official Ubuntu repositories that the workflow already
relies on.

The MinGW and macOS builds were out of scope, as the former doesn't run
tests and the latter is not defined here (and xvfb-run doesn't work on
Darwin).