Deprecate highlighting_is_*_style() in favor of editor_is_*_at()
highlighting_is_*_style() API cannot deal with substyles, because they
are inherently dynamic, but the API of these functions don't have
access to a mean of fetching the required dynamic info.

So, introduce replacement API editor_is_*_a(), and deprecate the other
one.

The new API takes an editor and a position, which allows to retrieve
the required dynamic information, and leaves room to further extend the
implementation if need arises.

Some design choices:

Q: Why use a GeanyEditor rather than a ScintillaObject?
A: We don't currently have original APIs that take a ScintillaObject.
   The APIs we have that do take one are mere wrapper around one or a
   few Scintilla calls, not original APIs altogether.  This means that
   one that do doesn't really have a spot to reside in, both in naming
   and source location.
   Additionally, this could allow future addition to the API that
   require extra data held on the GeanyEditor instance.
   The only downside is that this doesn't necessarily cover custom
   Scintilla instances, but it's a rather odd use case of having one
   and requiring this kind of info from it.

Q: Why replace style with position?  Current implementation would work
   fine with style as well.
A: Because of flexibility, and current usage.
   Flexibility: using a position could allow for looking around in the
   future if the exact meaning of a style could depend on the ones
   around, or if Scintilla had extra separate info that could be useful
   for determining the category that is not tied to the style itself.
   Reviewing current usage suggests that all callers perform a slight
   variation on `highlighting_is_*_style(lexer, style_at(sci, pos))`,
   so accepting a position directly just makes the API easier to use.
   The very few callers that use the style separately as well merely
   suffer from one additional call in this situation.
   Finally, there is no reason to tie the question "is there a comment
   here?" to a style from an user standpoint, so requiring one simply
   leak some internals when answering this question.