Fix Find in Files invalid directory error (#3818)
When the Find in Files path doesn't exist, the status bar says "Cannot execute grep tool...: Failed to change the working directory. Check the path setting in Preferences". Which is confusing and hides the real problem in the middle of irrelevant text. Instead, show a dialog with the correct error only. Fixes #1347.

* Fix FiF capitalization
* Set FIF search/directory entry background to indicate error