Support setting SCI_SETCARETPERIOD
To keep the things related to the caret at the same place, I used
filetypes.common.

In addition to allowing users modify the caret to their needs, this
can be used for debugging CPU usage as this should be the only timer
in Geany and when set to 0, Geany should (theoretically) use absolutely
no CPU.

When unset, it keeps using the system default value.

Fixes #3410.