Do not create empty file filters
The xdg-desktop-portal file chooser implementation (that can be forced
with `GTK_USE_PORTAL=1`) doesn't allow those and fails to create the
dialog if there are any such filter.

This seems like a potential xdg-desktop-portal bug, but it's easy
enough to workaround, and there's no guarantee yet it will change.

It's also unlikely to actually happen that we create such an empty
filter, but it is possible if a filetype doesn't have a corresponding
filetype_extensions.conf entry.