Avoid unwanted selection when infobar shows by mouse click in Scintilla
Steps to reporduce:
1. Open file A in Geany
2. In a separate window, edit the same file (using e.g. vi)
3. Click to Geany editor so it regains focus
4. An infobar appears. As it appears, several lines in editor are selected

The selection is caused by the fact that as the same time the editor
is being clicked, Geany checks if the document was modified and shows
the info bar. The info bar makes the Scintilla editor smaller and it
scrolls by the amount of lines corresponding to the height of the
infobar. All this happens in the click handler and for Scintilla it
appears as if the mouse button was presses during the scroll and it makes
the selection.

Showing infobars on idle seems to fix the problem.

Fixes #3906.