Also use GIO when using utils_write_file()
This patch unifies the behavior of utils_write_file() with
write_data_to_disk() so all 3 methods of file saving are supported.

From having a look at where this function is used, by Geany it's
used only to store configuration files in the .config directory and
some auxiliary functions like writing the export file by the exporter
plugin. In the geany-plugins project this function is used only for
saving configuration files.

Before this patch, by ignoring the default GIO file operation settings,
the fwrite() method was used by default for saving configuration files
which wasn't very safe and could lead to configuration file loss under
some extreme conditions.

The documentation of this function was changed to reflect the situation
better than now - the previous comment made the impression that file
permissions were always preserved with this method but it wasn't the
case and depended on the exact configuration. Technically this is an
API change but since this function is used for saving configuration
files only in practice, it should not be a problem.