Modify utils_write_file() to always use g_file_replace_contents()
From having a look at where this function is used, by Geany it's
used only to store configuration files in the .config directory and
some auxiliary functions like writing the export file by the exporter
plugin. In the geany-plugins project this function is used only for
saving configuration files.

Before this patch, the fwrite() method was used by default for saving
configuration files (because g_file_replace_contents() normally used
by Geany wasn't used in this function) which wasn't very safe and could
lead to configuration file loss under some extreme conditions.

The documentation of this function was changed to reflect the situation
better than now - the previous comment made the impression that file
permissions were always preserved with this method but it wasn't the
case and depended on the exact configuration.