Make "Find" the default button in the wrap search dialog
When the "Wrap search and find again?" dialog appears, it makes more
sense to wrap instead of cancel the dialog by default (the dialog
can always be canceled by Escape).

The implementation of this patch more or less duplicates the code of
dialogs_show_question_full() and makes the Find button default. While it
would be possible to add one more parameter to dialogs_show_question_full()
controlling this behavior, there's currently no other use of this in
Geany (in all other cases, the action button performs a destructive
action so Cancel is default) so I believe having this local implementation
of the dialog is a better option not complicating other uses.

Fixes #1192.