tagmanager: Unify Verilog and SystemVerilog tags
Use a consistent and unified map_* and group_* for Verilog and
SystemVerilog.

Some SystemVerilog features have been mirrored into Verilog,
resulting in the following changes to the latter:
- The following categories have been split:
    - constant <--> define
    - port <--> net/register
- The following categories have been merged:
    - event -><- net/register
- The following categories have been added (enabled):
    - block
    - instance
- Changes in group icons.

Note that module instantiation (in both Verilog and SystemVerilog) now
creates two ctags: one for a module definition and one for an instance.
The former should be removed, since the module isn't being defined here,
and Ctrl-click on the module name shouldn't take us here.
This needs to be fixed in ctags.