Split SystemVerilog keywords into type / not type
Moved all SystemVerilog keywords that are used when declaring signals
to a separate `word3` category, as it was done for Verilog in PR #4037.

The list has been obtained by picking all the keywords marked as
K_(EVENT|REGISTER|LOCALPARAM|PORT|PARAMETER|CONSTANT|SPECPARAM|NET)
in ctags/parsers/verilog.c, extended with additional keywords from
SystemVerilog-2017 section 6.7 "Net declarations" (as it was done for
Verilog in PR #4037), and with `signed` and `unsigned`.

The result is a superset of the `word3` category for Verilog,
extended with some keywords such as `logic`, `int`, `ref`, etc.