Add SystemVerilog ctags unit tests
Added two ctags unit tests for SystemVerilog: one with a "normal" design
(an executable testbench) to exercise logic and simulation constructs,
and one with a package and several OOP-related constructs.

It should be noted that some incorrect behavior has been found in ctags
when writing the tests; these were written down as comments in the test
files, and should be fixed in upstream ctags.