Fix GTK runtime errors generated by the plugin manager
The use of geany_wrap_label_new() together with the list below seems
to interact in a strange way somehow and generate lots of runtime
warnings/errors.

This is a (cowardly) attempt by fixing this without having to understand
what is going on ;-). Simply using gtk_label_new() seems to resolve
the issue.

Also, I think the text can be made a bit shorter in case the translation
is too wide for some languages. The "at startup" is kind of misleading
anyway since plugins are loaded immediately when checked in the box
below and since this is a preference, it's remembered so I think it's
pretty clear it happens at startup too.

Fixes #2065.