Drop homogenous spacing of additional buttons in search/replace dialog
In some languages, the text in the buttons of the search-replace dialog
is quite long. For instance, in Czech the "In session" button contains
"Ve všech otevřených dokumentech" (in all open documents). Homogenous
spacing of all the buttons means that all of them have the width of
the longest string which makes the search dialog extremely wide.

This patch drops the use of the GtkButtonBox which uses homogenous
spacing and uses ordinary GtkBox instead. However, to semi-simulate
the "homogenous" appearance of buttons, the patch sets the minimum
width of the buttons to 130 so all buttons whose text fits to this
width have the same size (for most languages I tried, nearly all
strings fit this width).

Note that the buttons in the action area of the dialog are still
sized homogenously as I didn't find a way to modify this behavior.
However, this doesn't seem to be a problem in practice with the
languages I tested.