Drop homogenous sizing for checkbox buttons
The hbox into which the 3+3 checkboxes with options are stored
is spaced homogenously right now. In some locales (German for instance)
some translation in the first column is much larger than all the
tranlsations in the second column which with the homogenous sizing
leads to a lot of wasted space in the second column and too
wide dialog.

This patch fixes that by dropping the homogenous spacing.