prefs: Edit the correct keybinding when there is a filter
When the keybinding list is filtered and the keybinding cell is edited,
store the value in the correct row instead of the one corresponding to
the Nth visible one, but in the unfiltered tree.

E.g., if filtering with `redo` it shows only 1 item, and editing it
should change that one (Edit/Redo), not the first of the unfiltered
list (File/New).

Related-to: #4192