Make sure that parents appear before their children when constructing symbol tree (#4063)
The symbol tree construction code assumes that tag parents appear before
their children. At the moment tag list is sorted by line numbers which
works for C structs

struct Foo {
    int bar;
    int baz;
}

where the tag Foo (which is a parent of bar and baz) is sorted before bar
and baz. However, this doesn't work when the name of the struct (or
a variable of an anonymous struct) follows the struct members such as
in the following SystemVerilog example:

struct packed {
    byte s_a, s_b, s_c;
    string s_str;
} s_member;

Here the members are sorted before the parent and the symbol tree
construction code doesn't work correctly.

This patch sorts symbols so that parents are sorted before their
members - only if tags aren't in a parent-child relationship the original
sorting by line is performed.

Fixes #4060.

Co-authored-by: Colomban Wendling <lists.ban@herbesfolles.org>