Fix build placeholder replacement
Fixes `%p` support that was broken in #4250, and avoids finding
placeholders in replacement strings that can lead to unexpected or
dangerous issues (e.g. if a path contains a placeholder).

The implementation is a little over-engineered, but might allow to
share the placeholder replacement logic with other code, as well as
decoupling string management and replacement generation.

Fixes #4314.