Remove position argument from ui_update_statusbar()
This function is always actually called with the document's current
position, and all but one caller pass `-1` already -- the only one
giving an explicit position gives `sci_get_current_position()`, which
behaves exactly the same as `-1`.

Thus, remove the unnecessary argument, which simplifies the calls and
will simplify an upcoming change.