Use utils_replace_placeholders() in ui_update_statusbar()
This moves the string management logic outside the placeholder
replacement logic, and fixes a corner case, when there is a % at the
end of the format string.

It also changes handling of unknown placeholders, keeping them entirely
unmodified instead of inserting only the character after the `%`.
Neither is necessarily perfect, but I don't think it's a regression or
a change that is relevant regarding compatibility.