Drop win32_open_browser()
On Windows, I get

```
21:15:27: Failed to open URI "C:/Program Files/Geany/share/doc/geany/html/index.html#general-startup-preferences": The system cannot find the file specified.
```

when pressing the Help button from the Preferences dialog. The problem
seems to be caused by the #suffix of the URI.

Using gtk_show_uri_on_window() instead of win32_open_browser() seems
to "work" so the web browser is shown - however, the #suffix is
removed so it doesn't navigate to the linked section in help. Still
better than not showing the help window at all.