Use Consolas as the default font on Windows
By default, the monospaced font on Windows is Courier New which isn't
exactly the nicest font for development because of the serifs (plus it
is rendered extremely thin on HiDPI displays). I expect that the first
thing everyone has to do on Windows is to change this font to something
else.

Since Windows Vista, Microsoft seems to ship a "modern" sans-serif
monospaced font called Consolas which looks as a much better font
for coding.

The rest of the patch just moves existing font configuration for other
platforms under a single ifdef/elif/else.